all are done

+++ components/Todo.tsx
@@ -4,6 +4,9 @@ import design from "../app/sass/todo.module.scss";
 // import { BiEdit } from "react-icons/bi";
 // import { BsTrash } from "react-icons/bs";
 import { useTodosContext } from "@/app/hooks/useTodoContext";
+import { useSession } from "next-auth/react";
+import { useRouter } from "next/navigation";
+
 interface TodoProps {
   todo: any;
 }
@@ -11,9 +14,15 @@ interface TodoProps {
 const Todo: React.FC<TodoProps> = ({ todo }) => {
   const { text, _id } = todo;
   const { dispatch } = useTodosContext();
+  const router = useRouter();
+  const { data: session } = useSession();
 
   /* HANDLE DELETE */
   const handleDelete = async (_id: string) => {
+    if (!session) {
+      router.push("/signin");
+      return;
+    }
     const res = await fetch(`/api/todos/${_id}`, {
       method: "DELETE",
     });
@@ -24,16 +33,24 @@ const Todo: React.FC<TodoProps> = ({ todo }) => {
     }
   };
 
+  /* handleEditBtn */
+  const handleEditBtn = (_id: string) => {
+    if (session) {
+      return router.push(`/edit-todo/${_id}`);
+    } else {
+      return router.push("/signin");
+    }
+  };
+
   return (
     <section className={design.section_container}>
       <div className={design.todo_div}>
         <h2>{text}</h2>
         <div className={design.todo_flex}>
-          <button className={design.editBtn}>
-            <Link className={design.link_design} href={`/edit-todo/${_id}`}>
-              edit
-            </Link>
+          <button onClick={() => handleEditBtn(_id)} className={design.editBtn}>
+            edit
           </button>
+
           <button
             onClick={() => handleDelete(_id)}
             className={design.deleteBtn}
